+++ Runtime/Scripts/Scenes/SceneManagerExtender.cs
@@ -9,54 +9,77 @@ using UnityEngine;
 
 using System;
 using System.Linq;
-using UnityEngine.SceneManagement;
 
+using UnityEngine.SceneManagement;
 using PolytopeSolutions.Toolset.GlobalTools.Types;
 
 namespace PolytopeSolutions.Toolset.Scenes {
     public class SceneManagerExtender : TManager<SceneManagerExtender> {
         [SerializeField] private string loaderSceneName;
         [SerializeField] private float smoothTransitionTime = 0.5f;
-        [SerializeField] private float minWaitTime=2f;
+        [SerializeField] private float minWaitTime = 2f;
         private Coroutine currentLoadingProcess;
         private float smoothProgress;
         public float SmoothProgress => this.smoothProgress;
         private float progress;
         public float Progress => this.progress;
+        private void UpdateProgress(float t) {
+            this.progress = t;
+        }
+
+        public string PreviousSceneName { get; private set; }
+        public string CurrentSceneName { get; private set; }
 
         public Scene activeScene => SceneManager.GetActiveScene();
         public string activeSceneName => this.activeScene.name;
 
-        public delegate void SceneEvent(string sceneName);
-        private Dictionary<string, Dictionary<string, SceneEvent>> OnSceneLoaded = new Dictionary<string, Dictionary<string, SceneEvent>>();
-        private Dictionary<string, Dictionary<string, SceneEvent>> OnSceneUnloaded = new Dictionary<string, Dictionary<string, SceneEvent>>();
-
-        public void RegisterSingletonOnSceneLoadedEvent(string sceneName, string source, SceneEvent sceneEvent) { 
-            if (!this.OnSceneLoaded.ContainsKey(sceneName))
-                this.OnSceneLoaded.Add(sceneName, new Dictionary<string, SceneEvent>());
-            if (!this.OnSceneLoaded[sceneName].ContainsKey(source))
-                this.OnSceneLoaded[sceneName].Add(source, sceneEvent);
+        public delegate IEnumerator SceneSetUpCoroutine(int currentOperationIndex, int operationCount, Action<float> updateProgress);
+        private Dictionary<string, Dictionary<string, SceneSetUpCoroutine>> OnSceneSetup = new Dictionary<string, Dictionary<string, SceneSetUpCoroutine>>();
+        private Dictionary<string, Dictionary<string, Action>> OnAfterSceneLoaded = new Dictionary<string, Dictionary<string, Action>>();
+        private Dictionary<string, Dictionary<string, Action>> OnBeforeSceneUnloaded = new Dictionary<string, Dictionary<string, Action>>();
+        
+        public void RegisterSingletonOnSceneSetupEvent(string sceneName, string source, SceneSetUpCoroutine sceneEvent) { 
+            if (!this.OnSceneSetup.ContainsKey(sceneName))
+                this.OnSceneSetup.Add(sceneName, new Dictionary<string, SceneSetUpCoroutine>());
+            if (!this.OnSceneSetup[sceneName].ContainsKey(source))
+                this.OnSceneSetup[sceneName].Add(source, sceneEvent);
             else
-                this.OnSceneLoaded[sceneName][source] = sceneEvent;
+                this.OnSceneSetup[sceneName][source] = sceneEvent;
         }
-        public void RegisterSingletonOnSceneUnloadedEvent(string sceneName, string source, SceneEvent sceneEvent) {
-            if (!this.OnSceneUnloaded.ContainsKey(sceneName))
-                this.OnSceneUnloaded.Add(sceneName, new Dictionary<string, SceneEvent>());
-
-            if (!this.OnSceneUnloaded[sceneName].ContainsKey(source))
-                this.OnSceneUnloaded[sceneName].Add(source, sceneEvent);
+        public void RegisterSingletonOnAfterSceneLoadedEvent(string sceneName, string source, Action sceneEvent) { 
+            if (!this.OnAfterSceneLoaded.ContainsKey(sceneName))
+                this.OnAfterSceneLoaded.Add(sceneName, new Dictionary<string, Action>());
+            if (!this.OnAfterSceneLoaded[sceneName].ContainsKey(source))
+                this.OnAfterSceneLoaded[sceneName].Add(source, sceneEvent);
             else
-                this.OnSceneUnloaded[sceneName][source] = sceneEvent;
+                this.OnAfterSceneLoaded[sceneName][source] = sceneEvent;
         }
-        public void UnregisterSingletonOnSceneLoadedEvent(string sceneName, string source, SceneEvent sceneEvent) {
-            if (this.OnSceneLoaded.ContainsKey(sceneName) && this.OnSceneLoaded[sceneName].ContainsKey(source))
-                this.OnSceneLoaded[sceneName][source] = null;
+        public void RegisterSingletonOnBeforeSceneUnloadedEvent(string sceneName, string source, Action sceneEvent) {
+            if (!this.OnBeforeSceneUnloaded.ContainsKey(sceneName))
+                this.OnBeforeSceneUnloaded.Add(sceneName, new Dictionary<string, Action>());
+
+            if (!this.OnBeforeSceneUnloaded[sceneName].ContainsKey(source))
+                this.OnBeforeSceneUnloaded[sceneName].Add(source, sceneEvent);
+            else
+                this.OnBeforeSceneUnloaded[sceneName][source] = sceneEvent;
         }
-        public void UnregisterSingletonOnSceneUnloadedEvent(string sceneName, string source, SceneEvent sceneEvent){
-            if (this.OnSceneUnloaded.ContainsKey(sceneName) && this.OnSceneUnloaded[sceneName].ContainsKey(source))
-                this.OnSceneUnloaded[sceneName][source] = null;
+        public void UnregisterSingletonOnSceneSetupEvent(string sceneName, string source) {
+            if (this.OnSceneSetup.ContainsKey(sceneName) && this.OnSceneSetup[sceneName].ContainsKey(source))
+                this.OnSceneSetup[sceneName].Remove(source);
+        }
+        public void UnregisterSingletonOnAfterSceneLoadedEvent(string sceneName, string source) {
+            if (this.OnAfterSceneLoaded.ContainsKey(sceneName) && this.OnAfterSceneLoaded[sceneName].ContainsKey(source))
+                this.OnAfterSceneLoaded[sceneName].Remove(source);
+        }
+        public void UnregisterSingletonOnBeforeSceneUnloadedEvent(string sceneName, string source){
+            if (this.OnBeforeSceneUnloaded.ContainsKey(sceneName) && this.OnBeforeSceneUnloaded[sceneName].ContainsKey(source))
+                this.OnBeforeSceneUnloaded[sceneName].Remove(source);
         }
 
+        public void Awake() { 
+            // through intance to ensuere it is assigned from the beginning.
+            SceneManagerExtender.instance.CurrentSceneName = this.activeSceneName;
+        }
         public void LoadScene(string sceneName, bool skipLoadingScene) {
             if (this.activeSceneName != sceneName) {
                 if (this.currentLoadingProcess != null) StopCoroutine(this.currentLoadingProcess);
@@ -64,59 +87,81 @@ namespace PolytopeSolutions.Toolset.Scenes {
             }
         }
         private IEnumerator LoadSceneCoroutine(string sceneName, bool skipLoadingScene) {
+            this.PreviousSceneName = this.CurrentSceneName;
+            this.CurrentSceneName = sceneName;
+
             skipLoadingScene &= !string.IsNullOrEmpty(this.loaderSceneName);
             DateTime startTime = DateTime.Now;
             this.progress = 0f;
-            string currentSceneName = this.activeSceneName;
+            int currentOperationIndex = 0;
+            List<Scene> currentScenes = GetActiveScenes();
+            int extraOperations = (this.OnSceneSetup.ContainsKey(sceneName)) ?
+                this.OnSceneSetup[sceneName].Count : 0;
+            int operationCount = 2              // Load New
+                + currentScenes.Count() * 2     // Unload Current, Unload Current Assets
+                + extraOperations               // Setup Operations if any
+                + ((skipLoadingScene) ? 0 : 2); // Load Loader, Unload Loader
             #if DEBUG
-            Debug.Log("SceneManagerExtender: Loading scene: " + sceneName + " from: " + currentSceneName + ". Skipping loadscene: " + skipLoadingScene);
+            Debug.Log("SceneManagerExtender: Loading scene: " + this.CurrentSceneName + " from: " + this.PreviousSceneName + ". Skipping loadscene: " + skipLoadingScene);
             #endif
             if (!skipLoadingScene) {
                 // 1. Load the loading scene if it was specified.
                 // - If there is loading scene - load it additively.
                 AsyncOperation loaderSceneLoad = SceneManager.LoadSceneAsync(this.loaderSceneName, LoadSceneMode.Additive);
-                yield return AwaitSceneAsyncOperation(loaderSceneLoad, Vector2.zero);
+                yield return AwaitSceneAsyncOperation(loaderSceneLoad,
+                    currentOperationIndex, operationCount);
+                currentOperationIndex++;
                 #if DEBUG2
                 Debug.Log("SceneManagerExtender: Loader scene loaded");
                 #endif
                 // - Trigger SmoothTransition coroutine.
-                yield return SmoothTransition(true);
+                yield return SmoothTransition(true, operationCount);
                 #if DEBUG2
                 Debug.Log("SceneManagerExtender: Loader smooth transitioning in complete");
                 #endif
-                // 2. Unload the old scene
-                // (NB! after loading scene as unloading can't happen without any scenes)
-                yield return UnloadOldScene(currentSceneName, skipLoadingScene);
-                #if DEBUG2
-                Debug.Log("SceneManagerExtender: Old scene unloaded");
-                #endif
+                // 2. Unload the old scenes
+                yield return UnloadOldScenes(currentScenes, currentOperationIndex, operationCount);
+                currentOperationIndex += currentScenes.Count() * 2;
             }
             // 3. Load the new scene
-            yield return LoadNewScene(sceneName, skipLoadingScene, startTime);
+            // - Launch new scene loading and pause it's activation until complete.
+            AsyncOperation newSceneLoad = SceneManager.LoadSceneAsync(sceneName, LoadSceneMode.Additive);
+            newSceneLoad.allowSceneActivation = false;
+            yield return LoadNewScene(newSceneLoad, sceneName, startTime,
+                currentOperationIndex, operationCount);
+            currentOperationIndex += 1 + extraOperations;
             #if DEBUG2
-            Debug.Log("SceneManagerExtender: New scene loaded. Active Scene: " + this.activeSceneName);
+            Debug.Log("SceneManagerExtender: New scene loaded.");
+            #endif
+            // - Wait for the remaining minimum wait time if relevant.
+            float waitTime = (skipLoadingScene) ? 0f :
+                Mathf.Max(0, this.minWaitTime - (float)(DateTime.Now - startTime).TotalSeconds - this.smoothTransitionTime * 2);
+            yield return new WaitForSeconds(waitTime);
+            // - Realease the scene loading async queue and Activate the scene.
+            yield return ActivateScene(newSceneLoad, sceneName);
+            #if DEBUG2
+            Debug.Log("SceneManagerExtender: New scene activated.");
             #endif
 
             if (!skipLoadingScene) {
                 // 4. Unload the loader scene if it was loaded.
                 // - Trigger SmoothTransition coroutine.
-                yield return SmoothTransition(false);
+                yield return SmoothTransition(false, operationCount);
                 #if DEBUG2
                 Debug.Log("SceneManagerExtender: Loader smooth transitioning out complete");
                 #endif
                 // - Unload the loading scene.
                 AsyncOperation loaderSceneUnload = SceneManager.UnloadSceneAsync(this.loaderSceneName);
-                yield return AwaitSceneAsyncOperation(loaderSceneUnload, Vector2.one);
+                yield return AwaitSceneAsyncOperation(loaderSceneUnload,
+                    currentOperationIndex, operationCount);
+                currentOperationIndex++;
                 #if DEBUG2
                 Debug.Log("SceneManagerExtender: Loader scene unloaded");
                 #endif
             } else {
-                // 5. Unload the old scene
-                // (NB! after as if loading scene is skipped unloading can't happen without any scenes)
-                yield return UnloadOldScene(currentSceneName, skipLoadingScene);
-                #if DEBUG2
-                Debug.Log("SceneManagerExtender: Old scene unloaded");
-                #endif
+                // 5. Unload the old scenes
+                yield return UnloadOldScenes(currentScenes, currentOperationIndex, operationCount);
+                currentOperationIndex += currentScenes.Count() * 2;
             }
             // 6. Reset loading coroutine.
             this.currentLoadingProcess = null;
@@ -126,52 +171,110 @@ namespace PolytopeSolutions.Toolset.Scenes {
                 + ". Active scene: " + this.activeSceneName);
             #endif
         }
-        private IEnumerator SmoothTransition(bool inTransition) { 
+        private List<Scene> GetActiveScenes() { 
+            List<Scene> scenes = new List<Scene>();
+            for (int i = 0; i < SceneManager.sceneCount; i++) {
+                Scene scene = SceneManager.GetSceneAt(i);
+                if (scene.isLoaded)
+                    scenes.Add(scene);
+            }
+            return scenes;
+        }
+        private IEnumerator SmoothTransition(bool inTransition, int operationsCount) { 
             this.smoothProgress = (inTransition) ? 0f : 1f;
             DateTime startTime = DateTime.Now;
-            float timePassed = 0f;
+            float timePassed = 0f, t = 0f;
             do {
                 yield return null;
                 timePassed = (float)(DateTime.Now - startTime).TotalSeconds;
-                this.smoothProgress = (inTransition) ? 
-                    Mathf.InverseLerp(0, this.smoothTransitionTime, timePassed) :
-                    Mathf.InverseLerp(this.smoothTransitionTime, 0, timePassed);
+                t = Mathf.InverseLerp(0, this.smoothTransitionTime, timePassed);
+                this.progress = (inTransition) ? 
+                    Mathf.Lerp(0f, 1f/operationsCount, t) :
+                    Mathf.Lerp(1f-1f/operationsCount, 1f, t);
+                this.smoothProgress = (inTransition) ? t : 1-t;
             } while (timePassed < this.smoothTransitionTime);
             this.smoothProgress = (inTransition) ? 1f : 0f;
+            this.progress = (inTransition) ? 1f/operationsCount : 1f;
             yield return null;
         }
-        private IEnumerator AwaitSceneAsyncOperation(AsyncOperation asyncOperation, Vector2 progressTargetRange) {
+        private IEnumerator AwaitSceneAsyncOperation(AsyncOperation asyncOperation,
+                int currentOperationIndex, int operationCount) {
+            Vector2 progressTargetRange = new Vector2(
+                Mathf.Lerp(0f, 1f, Mathf.InverseLerp(0, operationCount, currentOperationIndex)),
+                Mathf.Lerp(0f, 1f, Mathf.InverseLerp(0, operationCount, currentOperationIndex + 1))
+            );
+            this.progress = progressTargetRange.x;
             while (!asyncOperation.isDone && asyncOperation.progress < 0.9f) {
                 this.progress = Mathf.Lerp(progressTargetRange.x, progressTargetRange.y,
                     Mathf.InverseLerp(0f, 0.9f, asyncOperation.progress));
                 yield return null;
             }
+            this.progress = progressTargetRange.y;
         }
-        private IEnumerator LoadNewScene(string sceneName, bool skipLoadingScene, DateTime startTime) {
-            // - Launch new scene loading and pause it's activation until complete.
-            AsyncOperation newSceneLoad = SceneManager.LoadSceneAsync(sceneName, LoadSceneMode.Additive);
-            newSceneLoad.allowSceneActivation = false;
+        private IEnumerator LoadNewScene(AsyncOperation newSceneLoad, string sceneName, DateTime startTime,
+                int currentOperationIndex, int operationCount) {
             // - Wait until the next scene fully loads asynchronously.
-            Vector2 progressBarRange = (skipLoadingScene)?
-                new Vector2(0f, 0.5f) :
-                new Vector2(0.5f, 1f);
-            yield return AwaitSceneAsyncOperation(newSceneLoad, progressBarRange);
+            yield return AwaitSceneAsyncOperation(newSceneLoad, 
+                currentOperationIndex, operationCount);
+            currentOperationIndex++;
             // - Trigger OnScene Loaded events
-            if (this.OnSceneLoaded.ContainsKey(sceneName))
-                this.OnSceneLoaded[sceneName].Values.ToList().ForEach(item => item?.Invoke(sceneName));
-            this.progress = 1f;
-            // - Wait for the remaining minimum wait time if relevant.
-            float waitTime = (skipLoadingScene) ? 0f :
-                Mathf.Max(0, this.minWaitTime - (float)(DateTime.Now - startTime).TotalSeconds - this.smoothTransitionTime * 2);
-            yield return new WaitForSeconds(waitTime);
-            // - Realease the scene loading async queue.
-            newSceneLoad.allowSceneActivation = true;
-            // - Scene is loaded - set it as active
-            // (NB! has to happen after it is loaded, which won't happen till async queue is released)
+            if (this.OnAfterSceneLoaded.ContainsKey(sceneName))
+                this.OnAfterSceneLoaded[sceneName].Values.ToList().ForEach(item => item?.Invoke());
             yield return null;
+            // - Wait for setup events to finish
+            if (this.OnSceneSetup.ContainsKey(sceneName)) { 
+                foreach (KeyValuePair<string, SceneSetUpCoroutine> setupCoroutine in this.OnSceneSetup[sceneName]) { 
+                    #if DEBUG2
+                    Debug.Log("SceneManagerExtender: Executing setup coroutine: " + setupCoroutine.Key);
+                    #endif
+                    yield return setupCoroutine.Value(currentOperationIndex, operationCount, UpdateProgress);
+                    currentOperationIndex++;
+                }
+                #if DEBUG2
+                Debug.Log("SceneManagerExtender: Setup coroutines finished");
+                #endif
+            }
+        }
+        private void BeforeSceneUnload(string oldSceneName) {
+            // - Trigger OnBeforeSceneUnloaded events
+            if (this.OnBeforeSceneUnloaded.ContainsKey(oldSceneName))
+                this.OnBeforeSceneUnloaded[oldSceneName].Values.ToList().ForEach(item => item?.Invoke());
+        }
+        private IEnumerator UnloadOldScenes(List<Scene> currentScenes, int currentOperationIndex, int operationCount) {
+            foreach (Scene oldScene in currentScenes) { 
+                // - Trigger OnBeforeSceneUnloaded events
+                BeforeSceneUnload(oldScene.name);
+                // - Unload the old scene
+                // (NB! after loading loading scene as unloading can't happen without any scenes)
+                AsyncOperation oldSceneUnload = SceneManager.UnloadSceneAsync(oldScene.name);
+                oldSceneUnload.allowSceneActivation = false;
+                yield return UnloadOldScene(oldSceneUnload, oldScene.name, 
+                    currentOperationIndex, operationCount);
+                currentOperationIndex += 2;
+                #if DEBUG2
+                Debug.Log("SceneManagerExtender: Old scene ["+oldScene.name+"] unloaded");
+                #endif
+            }
+        }
+        private IEnumerator UnloadOldScene(AsyncOperation oldSceneUnload, string oldSceneName, 
+            int currentOperationIndex, int operationCount) {
+            // - Wait until the last scene fully unloads asynchronously.
+            yield return AwaitSceneAsyncOperation(oldSceneUnload, currentOperationIndex, operationCount);
+            currentOperationIndex++;
+            // In case of async unloading assets don't get released automatically. 
+            AsyncOperation unloadingAssets = Resources.UnloadUnusedAssets();
+            yield return AwaitSceneAsyncOperation(unloadingAssets, currentOperationIndex, operationCount);
+            // - Realease the scene loading async queue.
+            oldSceneUnload.allowSceneActivation = true;
+            yield return null;
+        }
+        private IEnumerator ActivateScene(AsyncOperation newSceneLoad, string sceneName) {
+            newSceneLoad.allowSceneActivation = true;
             bool loaded = false;
             Scene newScene;
             do {
+                if (this.activeSceneName == sceneName)
+                    break;
                 // Fix for build: Scene loading doesn't happen fast enough for the scene to be set as active.
                 // - wait and try to set the scene as active until it is set properly.
                 try{ 
@@ -184,21 +287,6 @@ namespace PolytopeSolutions.Toolset.Scenes {
                 yield return null;
             }
             while (!loaded);
-        }
-        private IEnumerator UnloadOldScene(string oldScene, bool skipLoadingScene) {
-            // - Trigger OnSceneUnloaded events
-            if (this.OnSceneUnloaded.ContainsKey(oldScene))
-                this.OnSceneUnloaded[oldScene].Values.ToList().ForEach(item => item?.Invoke(oldScene));
-            AsyncOperation oldSceneUnload = SceneManager.UnloadSceneAsync(oldScene);
-            oldSceneUnload.allowSceneActivation = false;
-            // - Wait until the last scene fully unloads asynchronously.
-            Vector2 progressBarRange = (skipLoadingScene) ?
-                new Vector2(0.5f, 1f) :
-                new Vector2(1f, 0.5f);
-            yield return AwaitSceneAsyncOperation(oldSceneUnload, progressBarRange);
-            // - Realease the scene loading async queue.
-            oldSceneUnload.allowSceneActivation = true;
-            yield return null;
         }
     }
 }